Rename project to openbao-template, use BAO_ vars
This replaces the name consul-template in many places,
switching to openbao-template. Note that non-code instances
of consul-template were not yet removed.

This also updates environment variables to use the new API
helpers (ReadBaoVariable) and to use the new BAO_ prefix
which will auto-detect for VAULT_ prefixes.

Technically this latter is a breaking change: the config of
openbao-template now outputs in BAO-prefixed variables
regardless of server to template against.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>